+++ CHANGELOG.md
@@ -1,3 +1,16 @@
 [1.0.0]
 
 Works great published version 1
+
+[1.4.0]
+
+A few bug fixes and more functionality added.
+
+Added textInputAction.
+
+## Braking Changes
+
+ TextField is set by default to call onSubmitted on a suggestion tap and also to clear the TextField on submit.
+
+ These can both be disabled with submitOnSuggestionTap and clearOnSubmit respectively.
+